chore: remove extra CardPageLayout file
